Refactor `Channel` and `Guild`
The structs (and functions to do with those, like parse and url) have
now been moved from `DiscordElixir.API` to `DiscordElixir.Model`.

This is in preparation for when we have models that have no
corresponding API. This would have led to some structs being found in
one group of modules and other structs being found in another group for
no good reason. This seems to be better code to me too!