Keepalive: always reset the timer on READY message
Before, it would only reset the timer if the heartbeat interval was
different to the one already stored. This means that if the websocket
process gets killed for some reason, the timer still attempts to send
the heartbeat message to the old websocket process and the websocket
fails to stay connected.